+++ src/Task/Task.php
@@ -62,6 +62,14 @@ class Task
     private $private = false;
 
     /**
+     * Shallow task will not print execution message/finish massage.
+     * Useful for success messages and info printing.
+     *
+     * @var bool
+     */
+    private $shallow = false;
+
+    /**
      * @param string $name Tasks name
      * @param callable $callback Task code
      */
@@ -248,5 +256,23 @@ class Task
     public function getAfter()
     {
         return $this->after;
+    }
+
+    /**
+     * Sets task shallow.
+     * @return $this
+     */
+    public function shallow()
+    {
+        $this->shallow = true;
+        return $this;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isShallow()
+    {
+        return $this->shallow;
     }
 }